Create multiple ts versions for t function and support const parameter
