+++ package.json
@@ -5,6 +5,13 @@
   "main": "./dist/cjs/i18next.js",
   "module": "./dist/esm/i18next.js",
   "types": "./index.d.ts",
+  "typesVersions": {
+    "<5.0": {
+      "typescript/t.d.ts": [
+        "typescript/t.v4.d.ts"
+      ]
+    }
+  },
   "exports": {
     "./package.json": "./package.json",
     ".": {
