+++ typescript/t.d.ts
@@ -1,0 +1,217 @@
+import type { $OmitArrayKeys, $PreservedValue, $Dictionary, $SpecialObject } from './helpers';
+import type { TypeOptions, Namespace, FlatNamespace, DefaultNamespace, TOptions } from './options';
+
+type $IsResourcesDefined = [keyof _Resources] extends [never] ? false : true;
+type $ValueIfResourcesDefined<Value, Fallback> = $IsResourcesDefined extends true
+  ? Value
+  : Fallback;
+type $FirstNamespace<Ns extends Namespace> = Ns extends readonly any[] ? Ns[0] : Ns;
+
+// Type Options
+type _ReturnObjects = TypeOptions['returnObjects'];
+type _ReturnNull = TypeOptions['returnNull'];
+type _KeySeparator = TypeOptions['keySeparator'];
+type _NsSeparator = TypeOptions['nsSeparator'];
+type _PluralSeparator = TypeOptions['pluralSeparator'];
+type _FallbackNamespace = TypeOptions['fallbackNS'];
+type _Resources = TypeOptions['resources'];
+type _JSONFormat = TypeOptions['jsonFormat'];
+type _InterpolationPrefix = TypeOptions['interpolationPrefix'];
+type _InterpolationSuffix = TypeOptions['interpolationSuffix'];
+
+type Resources = $ValueIfResourcesDefined<_Resources, $Dictionary<string>>;
+
+type PluralSuffix = 'zero' | 'one' | 'two' | 'few' | 'many' | 'other';
+
+type WithOrWithoutPlural<Key> = _JSONFormat extends 'v4'
+  ? Key extends `${infer KeyWithoutOrdinalPlural}${_PluralSeparator}ordinal${_PluralSeparator}${PluralSuffix}`
+    ? KeyWithoutOrdinalPlural | Key
+    : Key extends `${infer KeyWithoutPlural}${_PluralSeparator}${PluralSuffix}`
+    ? KeyWithoutPlural | Key
+    : Key
+  : Key;
+
+type JoinKeys<K1, K2> = `${K1 & string}${_KeySeparator}${K2 & string}`;
+type AppendNamespace<Ns, Keys> = `${Ns & string}${_NsSeparator}${Keys & string}`;
+
+/******************************************************
+ * Build all keys and key prefixes based on Resources *
+ ******************************************************/
+type KeysBuilderWithReturnObjects<Res, Key = keyof Res> = Key extends keyof Res
+  ? Res[Key] extends $Dictionary
+    ?
+        | JoinKeys<Key, WithOrWithoutPlural<keyof $OmitArrayKeys<Res[Key]>>>
+        | JoinKeys<Key, KeysBuilderWithReturnObjects<Res[Key]>>
+    : never
+  : never;
+
+type KeysBuilderWithoutReturnObjects<Res, Key = keyof $OmitArrayKeys<Res>> = Key extends keyof Res
+  ? Res[Key] extends $Dictionary
+    ? JoinKeys<Key, KeysBuilderWithoutReturnObjects<Res[Key]>>
+    : Key
+  : never;
+
+type KeysBuilder<Res, WithReturnObjects> = $IsResourcesDefined extends true
+  ? WithReturnObjects extends true
+    ? keyof Res | KeysBuilderWithReturnObjects<Res>
+    : KeysBuilderWithoutReturnObjects<Res>
+  : string;
+
+type KeysWithReturnObjects = {
+  [Ns in FlatNamespace]: WithOrWithoutPlural<KeysBuilder<Resources[Ns], true>>;
+};
+type KeysWithoutReturnObjects = {
+  [Ns in FlatNamespace]: WithOrWithoutPlural<KeysBuilder<Resources[Ns], false>>;
+};
+
+type ResourceKeys<WithReturnObjects = _ReturnObjects> = WithReturnObjects extends true
+  ? KeysWithReturnObjects
+  : KeysWithoutReturnObjects;
+
+/************************************************************************
+ * Parse t function keys based on the namespace, options and key prefix *
+ ************************************************************************/
+type KeysByTOptions<TOpt extends TOptions> = TOpt['returnObjects'] extends true
+  ? ResourceKeys<true>
+  : ResourceKeys;
+
+type NsByTOptions<Ns extends Namespace, TOpt extends TOptions> = TOpt['ns'] extends Namespace
+  ? TOpt['ns']
+  : Ns;
+
+type ParseKeysByKeyPrefix<Keys, KPrefix> = KPrefix extends string
+  ? Keys extends `${KPrefix}${_KeySeparator}${infer Key}`
+    ? Key
+    : never
+  : Keys;
+
+type ParseKeysByNamespaces<Ns extends Namespace, Keys> = Ns extends readonly (infer UnionNsps)[]
+  ? UnionNsps extends keyof Keys
+    ? AppendNamespace<UnionNsps, Keys[UnionNsps]>
+    : never
+  : never;
+
+type ParseKeysByFallbackNs<Keys extends $Dictionary> = _FallbackNamespace extends false
+  ? never
+  : _FallbackNamespace extends (infer UnionFallbackNs extends string)[]
+  ? Keys[UnionFallbackNs]
+  : Keys[_FallbackNamespace & string];
+
+type FilterKeysByContext<Keys, TOpt extends TOptions> = TOpt['context'] extends string
+  ? Keys extends `${infer Prefix}_${TOpt['context']}${infer Suffix}`
+    ? `${Prefix}${Suffix}`
+    : never
+  : Keys;
+
+export type ParseKeys<
+  Ns extends Namespace = DefaultNamespace,
+  TOpt extends TOptions = {},
+  KPrefix = undefined,
+  Keys extends $Dictionary = KeysByTOptions<TOpt>,
+  ActualNS extends Namespace = NsByTOptions<Ns, TOpt>,
+> = $IsResourcesDefined extends true
+  ? FilterKeysByContext<
+      | ParseKeysByKeyPrefix<Keys[$FirstNamespace<ActualNS>], KPrefix>
+      | ParseKeysByNamespaces<ActualNS, Keys>
+      | ParseKeysByFallbackNs<Keys>,
+      TOpt
+    >
+  : string;
+
+/*********************************************************
+ * Parse t function return type and interpolation values *
+ *********************************************************/
+type ParseInterpolationValues<Ret> =
+  Ret extends `${string}${_InterpolationPrefix}${infer Value}${_InterpolationSuffix}${infer Rest}`
+    ?
+        | (Value extends `${infer ActualValue},${string}` ? ActualValue : Value)
+        | ParseInterpolationValues<Rest>
+    : never;
+type InterpolationMap<Ret> = Record<$PreservedValue<ParseInterpolationValues<Ret>, string>, any>;
+
+type ParseTReturnPlural<
+  Res,
+  Key,
+  KeyWithPlural = `${Key & string}${_PluralSeparator}${PluralSuffix}`,
+  KeyWithOrdinalPlural = `${Key &
+    string}${_PluralSeparator}ordinal${_PluralSeparator}${PluralSuffix}`,
+> = Res[(KeyWithOrdinalPlural | KeyWithPlural | Key) & keyof Res];
+
+type ParseTReturn<Key, Res> = Key extends `${infer K1}${_KeySeparator}${infer RestKey}`
+  ? ParseTReturn<RestKey, Res[K1 & keyof Res]>
+  : ParseTReturnPlural<Res, Key>;
+
+type TReturnOptionalNull = _ReturnNull extends true ? null : never;
+type TReturnOptionalObjects<TOpt extends TOptions> = _ReturnObjects extends true
+  ? $SpecialObject | string
+  : TOpt['returnObjects'] extends true
+  ? $SpecialObject
+  : string;
+type DefaultTReturn<TOpt extends TOptions> = TReturnOptionalObjects<TOpt> | TReturnOptionalNull;
+
+type KeyWithContext<Key, TOpt extends TOptions> = TOpt['context'] extends string
+  ? `${Key & string}_${TOpt['context']}`
+  : Key;
+
+export type TFunctionReturn<
+  Ns extends Namespace,
+  Key,
+  TOpt extends TOptions,
+  ActualNS extends Namespace = NsByTOptions<Ns, TOpt>,
+  ActualKey = KeyWithContext<Key, TOpt>,
+> = $IsResourcesDefined extends true
+  ? ActualKey extends `${infer Nsp}${_NsSeparator}${infer RestKey}`
+    ? ParseTReturn<RestKey, Resources[Nsp & keyof Resources]>
+    : ParseTReturn<ActualKey, Resources[$FirstNamespace<ActualNS>]>
+  : DefaultTReturn<TOpt>;
+
+export type TFunctionDetailedResult<T = string> = {
+  /**
+   * The plain used key
+   */
+  usedKey: string;
+  /**
+   * The translation result.
+   */
+  res: T;
+  /**
+   * The key with context / plural
+   */
+  exactUsedKey: string;
+  /**
+   * The used language for this translation.
+   */
+  usedLng: string;
+  /**
+   * The used namespace for this translation.
+   */
+  usedNS: string;
+};
+
+type TFunctionReturnOptionalDetails<Ret, TOpt extends TOptions> = TOpt['returnDetails'] extends true
+  ? TFunctionDetailedResult<Ret>
+  : Ret;
+
+type AppendKeyPrefix<Key, KPrefix> = KPrefix extends string
+  ? `${KPrefix}${_KeySeparator}${Key & string}`
+  : Key;
+
+/**************************
+ * T function declaration *
+ **************************/
+export interface TFunction<Ns extends Namespace = DefaultNamespace, KPrefix = undefined> {
+  $TFunctionBrand: $IsResourcesDefined extends true ? `${$FirstNamespace<Ns>}` : never;
+  <
+    const Key extends ParseKeys<Ns, TOpt, KPrefix> | TemplateStringsArray,
+    const TOpt extends TOptions,
+    Ret extends TFunctionReturn<Ns, AppendKeyPrefix<Key, KPrefix>, TOpt>,
+    const ActualOptions extends TOpt & InterpolationMap<Ret> = TOpt & InterpolationMap<Ret>,
+  >(
+    ...args:
+      | [key: Key | Key[], options?: ActualOptions]
+      | [key: string | string[], options: TOpt & $Dictionary & { defaultValue: string }]
+      | [key: string | string[], defaultValue: string, options?: TOpt & $Dictionary]
+  ): TFunctionReturnOptionalDetails<Ret, TOpt>;
+}
+
+export type KeyPrefix<Ns extends Namespace> = ResourceKeys<true>[$FirstNamespace<Ns>] | undefined;
